add e2e test
