+++ assets/model_monitoring/components/tests/e2e/resources/mltable_iris_baseline_data_type_override/MLTable
@@ -1,0 +1,9 @@
+type: mltable
+
+paths:
+  - pattern: ./*.csv
+transformations:
+  - read_delimited:
+      delimiter: ','
+      encoding: ascii
+      header: all_files_same_headers
