+++ assets/model_monitoring/components/tests/e2e/resources/mltable_iris_baseline_data_type_override/iris_baseline.csv
@@ -1,0 +1,14 @@
+sepal_length,sepal_width,petal_length,petal_width,target
+5.1,1,1,4,setosa
+5.3,1,1,1,setosa
+2.5,1,1,2,setosa
+5.7,1,1,3,setosa
+4.4,1,1,4,setosa
+4.9,1,1,5,virginica
+5.0,1,1,6,versicolor
+4.4,5,5,5,virginica
+5.3,5,5,6,versicolor
+4.8,5,5,5,virginica
+5.1,5,5,6,versicolor
+4.9,5,5,5,virginica
+5.7,5,5,6,versicolor
