+++ assets/model_monitoring/components/tests/e2e/resources/mltable_iris_preprocessed_model_inputs_type_override/iris_baseline.csv
@@ -1,0 +1,6 @@
+sepal_length,sepal_width,petal_length,petal_width,target
+5.1,1,1,1,setosa
+5.2,2,5,2,setosa
+2.7,3,1,4,setosa
+3.8,4,5,4,setosa
+4.0,5,1,3,setosa
