+++ assets/model_monitoring/components/tests/e2e/utils/constants.py
@@ -82,6 +82,13 @@ DATA_ASSET_IRIS_PREPROCESSED_MODEL_INPUTS_NO_COMMON_COLUMNS = (
     "azureml:mltable_iris_preprocessed_model_inputs_no_common_columns:1"
 )
 
+DATA_ASSET_IRIS_PREPROCESSED_MODEL_INPUTS_TYPE_OVERRIDE = (
+    "azureml:mltable_iris_preprocessed_model_inputs_type_override:1"
+)
+DATA_ASSET_IRIS_BASELINE_DATA_TYPE_OVERRIDE = (
+    "azureml:mltable_iris_baseline_data_type_override:1"
+)
+
 # used for checking against histogram regressions where a numerical data-column has a single distinct value
 DATA_ASSET_IRIS_PREPROCESSED_MODEL_INPUTS_INT_SINGLE_VALUE_HISTOGRAM = (
     "azureml:mltable_iris_preprocessed_model_inputs_int_single_value_histogram:1"
